Split custody handlers by workflow
